수정
- ui 리소스 수정
- 스킬 포물선
- 아이템 생성 위치 변경,
- 아이템 콜라이더 크기 수정
- 업그레이드 패널 int -> float 형태로 변경

+++ Assets/3. Scripts/Upgrade/UpgradeSystem.cs
@@ -89,7 +89,7 @@ public class UpgradeSystem : MonoBehaviour
             var row = dataset[index];
 
             int grade = Convert.ToInt32(row["Grade"]);
-            int value = Convert.ToInt32(row["Value"]);
+            float value =  (float)Convert.ToDouble(row["Value"]);
 
             upgradeMenuItem.Construct(i, grade, value);
             upgradeMenuItem.EnableVfx(false);
@@ -153,7 +153,7 @@ else  // F (10% 추가, 총 100%)
 
             var row = dataset[typeOffset + gradeOffset];
             int grade = Convert.ToInt32(row["Grade"]);
-            int value = Convert.ToInt32(row["Value"]);
+            float value =  (float)Convert.ToDouble(row["Value"]);
 
             upgradeMenuItem.Construct(i, grade, value);
             upgradeMenuItem.EnableVfx(false);
